refactor(cli): split the command handlers into commands.rs
cli.rs held both the clap surface and the implementations behind it.
The handlers (and their tests) move to src/commands.rs; cli.rs keeps
the argument definitions, the dispatch predicates and parse_idle.
